Move the query system to rustc_query_impl.
