+++ compiler/rustc_middle/src/ty/context.rs
@@ -966,7 +966,7 @@ pub struct GlobalCtxt<'tcx> {
     /// Do not access this directly. It is only meant to be used by
     /// `DepGraph::try_mark_green()` and the query infrastructure.
     /// This is `None` if we are not incremental compilation mode
-    pub(crate) on_disk_cache: Option<OnDiskCache<'tcx>>,
+    pub on_disk_cache: Option<OnDiskCache<'tcx>>,
 
     pub queries: &'tcx dyn QueryEngine<'tcx>,
