+++ compiler/rustc_query_impl/src/profiling_support.rs
@@ -1,11 +1,10 @@
-use crate::ty::context::TyCtxt;
-use crate::ty::query::Query;
-use crate::ty::WithOptConstParam;
+use super::Query;
 use measureme::{StringComponent, StringId};
 use rustc_data_structures::fx::FxHashMap;
 use rustc_data_structures::profiling::SelfProfiler;
 use rustc_hir::def_id::{CrateNum, DefId, DefIndex, LocalDefId, CRATE_DEF_INDEX, LOCAL_CRATE};
 use rustc_hir::definitions::DefPathData;
+use rustc_middle::ty::{TyCtxt, WithOptConstParam};
 use rustc_query_system::query::{QueryCache, QueryState};
 use std::fmt::Debug;
 use std::io::Write;
